requirements.txt
streamlit==1.54.0
anthropic==0.79.0
openai==2.20.0
google-genai==1.63.0
python-dotenv==1.2.1
pandas==2.3.3
streamlit-local-storage==0.0.25
python-docx==1.1.2
pymupdf==1.25.4
